* edit relation - перенос
